pyproject.toml
[project]
name = "offline-scouting-manager"
version = "0.1.0"
description = "A Flask-based application for managing offline scouting data and operations for robotics outreach teams"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pyyaml>=6.0.3",
]
